# docs/api/admin.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "admin-service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/proposals": {
      "get": {
        "tags": [
          "proposals"
        ],
        "summary": "List Proposals",
        "description": "Paginated proposals listing (newest first).",
        "operationId": "list_proposals_admin_proposals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "proposed_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proposed By"
            }
          },
          {
            "name": "target",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return rows with created_at < this value (older page).",
              "title": "Cursor"
            },
            "description": "Return rows with created_at < this value (older page)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/proposals/{proposal_id}": {
      "get": {
        "tags": [
          "proposals"
        ],
        "summary": "Get Proposal",
        "description": "Return one proposal with full payload.",
        "operationId": "get_proposal_admin_proposals__proposal_id__get",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/proposals/{proposal_id}/approve": {
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Approve Proposal",
        "description": "Transition a proposal to ``approved`` and request a validation backtest.",
        "operationId": "approve_proposal_admin_proposals__proposal_id__approve_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveProposalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveProposalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/proposals/{proposal_id}/reject": {
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Reject Proposal",
        "description": "Transition a proposal to ``rejected`` with a required ``review_notes``.",
        "operationId": "reject_proposal_admin_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectProposalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectProposalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveProposalRequest": {
        "properties": {
          "review_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          },
          "skip_backtest": {
            "type": "boolean",
            "title": "Skip Backtest",
            "default": false
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Id"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "universe": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Universe",
            "default": "sp500"
          },
          "git_sha": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Sha"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApproveProposalRequest",
        "description": "``POST /admin/proposals/{id}/approve`` body.\n\nBy default, approval triggers a validation backtest. Operators can pass\n``skip_backtest=true`` for proposals whose ``target`` is not a strategy."
      },
      "ApproveProposalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewed_by": {
            "type": "string",
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          },
          "validation_backtest_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Backtest Id"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "reviewed_by",
          "reviewed_at"
        ],
        "title": "ApproveProposalResponse",
        "description": "Proposal row after approval."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProposalDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "proposed_by": {
            "type": "string",
            "title": "Proposed By"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          },
          "validation_backtest_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Backtest Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "current_value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Current Value"
          },
          "proposed_value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Proposed Value"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "estimated_impact": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Impact"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "applied_commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Commit Sha"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "proposed_by",
          "category",
          "target",
          "rationale",
          "status",
          "created_at"
        ],
        "title": "ProposalDetail",
        "description": "``GET /admin/proposals/{id}`` payload (full row)."
      },
...
      "ProposalsListResponse": {
        "properties": {
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/ProposalSummary"
            },
            "type": "array",
            "title": "Proposals"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as ``cursor`` query param (older ``created_at``) to fetch the next page."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "proposals",
          "limit"
        ],
        "title": "ProposalsListResponse",
        "description": "``GET /admin/proposals`` payload."
      },
...
      "RejectProposalRequest": {
        "properties": {
          "review_notes": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Review Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review_notes"
        ],
        "title": "RejectProposalRequest",
        "description": "``POST /admin/proposals/{id}/reject`` body."
      },
...
    }
  }
}